refactored code and updated loading animation
