+++ app/pages/_app.tsx
@@ -7,12 +7,8 @@ import type { Dispatch, SetStateAction } from "react";
 import io, { Socket } from "socket.io-client";
 import { socketEvent } from "../utils/socketServerHandler";
 import { v4 as uuidv4 } from "uuid";
-import {
-  EmisionUnits,
-  Game,
-  RegionDetails,
-  user as userType,
-} from "../utils/types/game";
+import { Game, user as userType } from "../utils/game";
+import { RegionDetails } from "../utils/details";
 import { Region } from "@prisma/client";
 
 import { ThemeProvider, createTheme } from "@mui/material/styles";
@@ -96,7 +92,7 @@ export default function App({ Component, pageProps }: AppProps) {
           ...user,
           gameCode: code,
           region: region,
-          emission: EmisionUnits[region],
+          emission: RegionDetails[region].emissionUnits,
         });
         router.push("/lobby");
       });
