+++ app/pages/index.tsx
@@ -9,7 +9,8 @@ import Tooltip from "@mui/material/Tooltip";
 
 import { socket, snackbarProps } from "./_app";
 import { socketEvent } from "../utils/socketServerHandler";
-import { DescriptiveTooltips, userState } from "../utils/types/game";
+import { DescriptiveTooltips } from "../utils/details";
+import { userState } from "../utils/game";
 import { gameCodeLength } from "../utils/constants";
 
 export default function Home({
