+++ app/pages/test/loading.tsx
@@ -1,0 +1,5 @@
+import Loading from "../../components/loading";
+
+export default function LoadingTest() {
+  return <Loading />;
+}
